moving core libraries of validator to separate classes
